Fix remaining unnecessary_cast clippy lint in av1 frame.rs
CI's clippy (1.97.0) flags `(read_f8(...)? + 3) as u8` as a redundant
cast (read_f8 already returns u8) where the locally available clippy
(1.94.0) did not. Drop the cast.

+++ tpt-kinetix-av1/src/frame.rs
@@ -563,7 +563,7 @@ impl FrameHeader {
 
         // --- CDEF ---
         let cdef_damping = if !lossless {
-            (read_f8(&mut br, 2)? + 3) as u8
+            read_f8(&mut br, 2)? + 3
         } else {
             0
         };
